+++ javascript/core/src/engine/util/sets.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import { container } from '../../factories.js'
+import { container } from '../../container.js'
 import type {
 	SetOperationArgs,
 	Step,
