+++ javascript/core/src/engine/verbs/dedupe.ts
@@ -4,8 +4,8 @@
  */
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 
-import { makeStepFunction, makeStepNode } from '../../factories.js'
 import type { DedupeArgs } from '../../types.js'
+import { makeStepFunction, makeStepNode } from '../factories.js'
 
 export const dedupe = makeStepFunction(doDedupe)
 export const dedupeNode = makeStepNode(doDedupe)
