+++ javascript/core/src/engine/verbs/derive.ts
@@ -6,9 +6,9 @@ import { escape } from 'arquero'
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 import type { ExprObject } from 'arquero/dist/types/table/transformable'
 
-import { makeStepFunction, makeStepNode } from '../../factories.js'
 import type { DeriveArgs } from '../../types.js'
 import { MathOperator } from '../../types.js'
+import { makeStepFunction, makeStepNode } from '../factories.js'
 
 export const derive = makeStepFunction(doDerive)
 export const deriveNode = makeStepNode(doDerive)
