+++ javascript/core/src/engine/verbs/fold.ts
@@ -5,8 +5,8 @@
 
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 
-import { makeStepFunction, makeStepNode } from '../../factories.js'
 import type { FoldArgs } from '../../types.js'
+import { makeStepFunction, makeStepNode } from '../factories.js'
 
 export const fold = makeStepFunction(doFold)
 export const foldNode = makeStepNode(doFold)
