+++ javascript/core/src/engine/verbs/groupby.ts
@@ -4,8 +4,8 @@
  */
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 
-import { makeStepFunction, makeStepNode } from '../../factories.js'
 import type { GroupbyArgs } from '../../types.js'
+import { makeStepFunction, makeStepNode } from '../factories.js'
 
 export const groupby = makeStepFunction(doGroupby)
 export const groupbyNode = makeStepNode(doGroupby)
