+++ javascript/core/src/engine/verbs/lookup.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import { container } from '../../factories.js'
+import { container } from '../../container.js'
 import type { TableStore } from '../../index.js'
 import type { LookupStep, TableContainer } from '../../types.js'
