+++ javascript/core/src/engine/verbs/pivot.ts
@@ -4,8 +4,8 @@
  */
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 
-import { makeStepFunction, makeStepNode } from '../../factories.js'
 import type { PivotArgs } from '../../types.js'
+import { makeStepFunction, makeStepNode } from '../factories.js'
 import { singleExpression } from '../util/index.js'
 
 export const pivot = makeStepFunction(doPivot)
