+++ javascript/core/src/engine/verbs/rename.ts
@@ -4,8 +4,8 @@
  */
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 
-import { makeStepFunction, makeStepNode } from '../../factories.js'
 import type { RenameArgs } from '../../types.js'
+import { makeStepFunction, makeStepNode } from '../factories.js'
 
 export const rename = makeStepFunction(doRename)
 export const renameNode = makeStepNode(doRename)
