+++ javascript/core/src/engine/verbs/select.ts
@@ -5,8 +5,8 @@
 import { all } from 'arquero'
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 
-import { makeStepFunction, makeStepNode } from '../../factories.js'
 import type { SelectArgs } from '../../types.js'
+import { makeStepFunction, makeStepNode } from '../factories.js'
 
 export const select = makeStepFunction(doSelect)
 export const selectNode = makeStepNode(doSelect)
