+++ javascript/core/src/engine/verbs/unfold.ts
@@ -6,8 +6,8 @@ import { from } from 'arquero'
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 import type { RowObject } from 'arquero/dist/types/table/table'
 
-import { makeStepFunction, makeStepNode } from '../../factories.js'
 import type { UnfoldArgs } from '../../types.js'
+import { makeStepFunction, makeStepNode } from '../factories.js'
 
 export const unfold = makeStepFunction(doUnfold)
 export const unfoldNode = makeStepNode(doUnfold)
