+++ javascript/core/src/engine/verbs/ungroup.ts
@@ -5,7 +5,7 @@
 
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 
-import { makeStepFunction, makeStepNode } from '../../factories.js'
+import { makeStepFunction, makeStepNode } from '../factories.js'
 
 export const ungroup = makeStepFunction(doUngroup)
 export const ungroupNode = makeStepNode(doUngroup)
