+++ javascript/core/src/engine/verbs/unorder.ts
@@ -4,7 +4,7 @@
  */
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 
-import { makeStepFunction, makeStepNode } from '../../factories.js'
+import { makeStepFunction, makeStepNode } from '../factories.js'
 
 export const unorder = makeStepFunction(doUnorder)
 export const unorderNode = makeStepNode(doUnorder)
