+++ javascript/dataflow-graph/src/__tests__/NodeImpl.spec.ts
